feature(dynamic-routes): consolidated the MemoryRouterProvider implementation, using a factory abstraction
