+++ MemoryRouterProvider/next-10/package.json
@@ -1,3 +1,3 @@
 {
-  "main": "../../dist/MemoryRouterProvider/MemoryRouterProvider-10"
+  "main": "../../dist/MemoryRouterProvider/next-10"
 }
