+++ MemoryRouterProvider/next-12/package.json
@@ -1,3 +1,3 @@
 {
-  "main": "../../dist/MemoryRouterProvider/MemoryRouterProvider-11.1"
+  "main": "../../dist/MemoryRouterProvider/next-11"
 }
